+++ main.tf
@@ -402,26 +402,29 @@ resource "aws_acm_certificate_validation" "alb_cert" {
 
 # ALB short names and ALBs target groups names
 locals {
-  alb_namespace_short          = substr(var.namespace, 0, 4)
-  alb_stage_short              = substr(var.stage, 0, 1)
-  alb_internal_name_short      = "${substr(var.name, 0, min(length(var.name), 18))}-i"
-  alb_external_name_short      = "${substr(var.name, 0, min(length(var.name), 18))}-e"
-  internal_alb_default_tg_name = "${local.alb_namespace_short}-${local.alb_stage_short}-${local.alb_internal_name_short}dtg"
-  external_alb_default_tg_name = "${local.alb_namespace_short}-${local.alb_stage_short}-${local.alb_external_name_short}dtg"
+  alb_namespace_short           = substr(var.namespace, 0, 4)
+  alb_stage_short               = substr(var.stage, 0, 1)
+  alb_internal_name_short       = "${substr(var.name, 0, min(length(var.name), 18))}-i"
+  alb_external_name_short       = "${substr(var.name, 0, min(length(var.name), 18))}-e"
+  alb_internal_default_tg_name  = "${local.alb_namespace_short}-${local.alb_stage_short}-${local.alb_internal_name_short}dtg"
+  alb_external_default_tg_name  = "${local.alb_namespace_short}-${local.alb_stage_short}-${local.alb_external_name_short}dtg"
 }
 
 module "alb_default_internal" {
-  source                                  = "git::https://github.com/cloudposse/terraform-aws-alb.git?ref=tags/0.24.0"
+  source                                  = "git::https://github.com/cloudposse/terraform-aws-alb.git?ref=tags/0.29.4"
   enabled                                 = var.alb_internal_enabled
   namespace                               = local.alb_namespace_short
   name                                    = local.alb_internal_name_short
   stage                                   = local.alb_stage_short
   attributes                              = var.attributes
   vpc_id                                  = var.vpc_id
-  security_group_ids                      = []
+  security_group_enabled                  = var.alb_internal_default_security_group_enabled
+  security_group_ids                      = var.alb_internal_additional_security_groups_list
+  http_ingress_cidr_blocks                = var.alb_internal_default_security_group_ingress_cidrs_blocks
+  https_ingress_cidr_blocks               = var.alb_internal_default_security_group_ingress_cidrs_blocks
   subnet_ids                              = var.private_subnets
   internal                                = true
-  target_group_name                       = local.internal_alb_default_tg_name
+  target_group_name                       = local.alb_internal_default_tg_name
   http_enabled                            = var.alb_internal_http_enable && var.alb_internal_enabled ? true : false
   http_redirect                           = var.alb_internal_http_redirect && var.alb_internal_http_enable && var.alb_internal_enabled ? true : false
   https_enabled                           = var.alb_internal_https_enable && var.alb_internal_enabled ? true : false
@@ -437,17 +440,20 @@ module "alb_default_internal" {
 }
 
 module "alb_default_external" {
-  source                                  = "git::https://github.com/cloudposse/terraform-aws-alb.git?ref=tags/0.24.0"
+  source                                  = "git::https://github.com/cloudposse/terraform-aws-alb.git?ref=tags/0.29.4"
   enabled                                 = var.alb_external_enabled
   namespace                               = local.alb_namespace_short
   name                                    = local.alb_external_name_short
   stage                                   = local.alb_stage_short
   attributes                              = var.attributes
   vpc_id                                  = var.vpc_id
-  security_group_ids                      = []
+  security_group_enabled                  = var.alb_external_default_security_group_enabled
+  security_group_ids                      = var.alb_external_additional_security_groups_list
+  http_ingress_cidr_blocks                = var.alb_external_default_security_group_ingress_cidrs_blocks
+  https_ingress_cidr_blocks               = var.alb_external_default_security_group_ingress_cidrs_blocks
   subnet_ids                              = var.public_subnets
   internal                                = false
-  target_group_name                       = local.external_alb_default_tg_name
+  target_group_name                       = local.alb_external_default_tg_name
   http_enabled                            = var.alb_external_http_enable && var.alb_external_enabled ? true : false
   http_redirect                           = var.alb_external_http_redirect && var.alb_external_http_enable && var.alb_external_enabled ? true : false
   https_enabled                           = var.alb_external_https_enable && var.alb_external_enabled ? true : false